# MigraTeck/apps/auth-api/openapi.yaml
openapi: "3.1.0"
info:
  title: MigraAuth API
  version: "1.0.0"
  description: >
    Centralized identity platform for the MigraTeck ecosystem.
    OAuth 2.1 + PKCE, session-based auth, MFA (TOTP), OIDC discovery.
  contact:
    name: MigraTeck
    url: https://migrateck.com
servers:
  - url: https://auth.migrateck.com
    description: Production
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /authorize:
    get:
      tags: [OAuth]
      operationId: authorize
      summary: Authorization endpoint (redirect-based)
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            enum: [S256]
        - name: scope
          in: query
          schema:
            type: string
            default: openid
        - name: nonce
          in: query
          schema:
            type: string
        - name: prompt
          in: query
          schema:
            type: string
            enum: [none, login, consent]
        - name: login_hint
          in: query
          schema:
            type: string
        - name: return_to
          in: query
          schema:
            type: string
            format: uri
      responses:
        "302":
          description: Redirect to login or back to client with auth code
        "400":
          $ref: "#/components/responses/BadRequest"
  /authorize/complete:
    post:
      tags: [OAuth]
      operationId: authorizeComplete
      summary: Issue auth code after login (called by auth-web)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorizeCompleteRequest"
      responses:
        "200":
          description: Auth code issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizeCompleteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /token:
    post:
      tags: [OAuth]
      operationId: exchangeToken
      summary: Token exchange (authorization_code or refresh_token)
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/TokenRequest"
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      responses:
        "200":
          description: Token set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /revoke:
    post:
      tags: [OAuth]
      operationId: revokeToken
      summary: Revoke a refresh token (RFC 7009)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeRequest"
      responses:
        "200":
          description: Token revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericRevokedResponse"
  /userinfo:
    get:
      tags: [OAuth]
      operationId: getUserInfo
      summary: OIDC UserInfo
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: User claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoResponse"

  # ─── Account ───────────────────────────────────────────────────
...
  /v1/login:
    post:
      tags: [Account]
      operationId: login
      summary: Authenticate with email or phone plus password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/refresh:
    post:
      tags: [Account]
      operationId: refresh
      summary: Rotate the first-party refresh token and session secret
      responses:
        "200":
          description: Session refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/mfa/totp/enroll:
    post:
      tags: [MFA]
      operationId: enrollTotp
      summary: Start TOTP enrollment
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Enrollment started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TotpEnrollResponse"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/mfa/totp/verify:
    post:
      tags: [MFA]
      operationId: verifyTotp
      summary: Verify a TOTP code (enrollment confirmation or login MFA)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TotpVerifyRequest"
      responses:
        "200":
          description: TOTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TotpVerifyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/mfa/disable:
    post:
      tags: [MFA]
      operationId: disableMfa
      summary: Disable MFA (requires password confirmation)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MfaDisableRequest"
      responses:
        "200":
          description: MFA disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericSuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── Sessions ──────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: migraauth_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: invalid_credentials
            message:
              type: string
              example: Invalid email or password.

    # ── Generic responses ──────────────────────────────────────
    GenericSuccessResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
    GenericRevokedResponse:
      type: object
      required: [revoked]
      properties:
        revoked:
          type: boolean
          example: true

    # ── Account schemas ────────────────────────────────────────
...
    LoginRequest:
      type: object
      required: [identifier, password, client_id]
      properties:
        identifier:
          type: string
          description: Email address or E.164 phone number
        password:
          type: string
        client_id:
          type: string
    LoginSuccessResponse:
      type: object
      required: [authenticated, requires_mfa, user, session]
      properties:
        authenticated:
          type: boolean
          enum: [true]
        requires_mfa:
          type: boolean
          enum: [false]
        user:
          $ref: "#/components/schemas/UserSummary"
        session:
          $ref: "#/components/schemas/SessionSummary"
    LoginVerificationRequiredResponse:
      type: object
      required: [status, challenge_id, channel, masked_destination, message]
      properties:
        status:
          type: string
          enum: [verification_required]
        challenge_id:
          type: string
          format: uuid
        channel:
          type: string
          enum: [email, sms]
        masked_destination:
          type: string
        message:
          type: string
    LoginMfaResponse:
      type: object
      required: [authenticated, requires_mfa]
      properties:
        authenticated:
          type: boolean
          enum: [false]
        requires_mfa:
          type: boolean
          enum: [true]
    LoginResponse:
      oneOf:
        - $ref: "#/components/schemas/LoginSuccessResponse"
        - $ref: "#/components/schemas/LoginVerificationRequiredResponse"
        - $ref: "#/components/schemas/LoginMfaResponse"
    RefreshResponse:
      type: object
      required: [authenticated, access_token, token_type, expires_in, user, session]
      properties:
        authenticated:
          type: boolean
          enum: [true]
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
        user:
          $ref: "#/components/schemas/UserSummary"
        session:
          $ref: "#/components/schemas/SessionSummary"
...
    UserSummary:
      type: object
      required: [id, email, phone_e164, status, email_verified, phone_verified]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          nullable: true
        phone_e164:
          type: string
          nullable: true
        status:
          type: string
        email_verified:
          type: boolean
        phone_verified:
          type: boolean
        display_name:
          type: string
    SessionSummary:
      type: object
      required: [id, created_at, expires_at]
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        ip_address:
          type: string
          nullable: true
        user_agent:
          type: string
          nullable: true

    # ── OAuth schemas ──────────────────────────────────────────
    AuthorizeCompleteRequest:
      type: object
      required: [client_id, redirect_uri, state, code_challenge, code_challenge_method]
      properties:
        client_id:
          type: string
        redirect_uri:
          type: string
          format: uri
        state:
          type: string
        code_challenge:
          type: string
        code_challenge_method:
          type: string
          enum: [S256]
        scope:
          type: string
        nonce:
          type: string
    AuthorizeCompleteResponse:
      type: object
      required: [redirect_uri, code, state]
      properties:
        redirect_uri:
          type: string
        code:
          type: string
        state:
          type: string
    TokenRequest:
      type: object
      required: [grant_type, client_id]
      properties:
        grant_type:
          type: string
          enum: [authorization_code, refresh_token]
        client_id:
          type: string
        code:
          type: string
        code_verifier:
          type: string
        redirect_uri:
          type: string
        refresh_token:
          type: string
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
        refresh_token:
          type: string
        id_token:
          type: string
        scope:
          type: string
    RevokeRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
    UserInfoResponse:
      type: object
      properties:
        sub:
          type: string
          format: uuid
        email:
          type: string
        email_verified:
          type: boolean
        name:
          type: string
        given_name:
          type: string
        family_name:
          type: string
        picture:
          type: string
        locale:
          type: string

    # ── MFA schemas ────────────────────────────────────────────
    TotpEnrollResponse:
      type: object
      required: [challenge_id, secret, otpauth_uri, recovery_codes]
      properties:
        challenge_id:
          type: string
          format: uuid
        secret:
          type: string
        otpauth_uri:
          type: string
        recovery_codes:
          type: array
          items:
            type: string
    TotpVerifyRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 6
        challenge_id:
          type: string
          format: uuid
    TotpVerifyResponse:
      type: object
      required: [verified]
      properties:
        verified:
          type: boolean
    MfaDisableRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string

    # ── Session schemas ────────────────────────────────────────
...